refactor(credit-check): decompose CreditCheckService into Crud/AiAnalysis/Risk/Override services
Behavior-preserving. Facade keeps the full public surface + 2-arg ctor + delegates;
internally constructs the 4 sub-services (keeps all 7 `new CreditCheckService(...)`
spec sites + module untouched). 2 override $transaction moved WHOLE; background
auto-score fire-and-forget + getAnthropicClient memoization preserved. tests green,
tsc 0, lint 0.

Sub-services exposed as public readonly fields (risk/ai/crud/override_) so the 5
specs that spied on a now-cross-service helper (performAIAnalysis, getAnthropicClient,
performClaudeAnalysis, performRuleBasedAnalysis, getCustomerHistory, calculateRiskScore)
retarget the spy to the owning sub-service instance. Spy-target only — NO assertion
changes. The other 3 specs are 100% untouched.
